message class in 5 apps

+++ Controller/Category.php
@@ -3,7 +3,7 @@ require_once 'Controller/Core/Action.php';
 require_once 'Model/Category.php';
 require_once 'Model/Core/Url.php';
 
-require_once 'Model/Core/Message.php';
+require_once 'Model/Core/Message.php';  
 
 
 class Controller_Category extends Contoller_Core_Action{
@@ -42,14 +42,22 @@ class Controller_Category extends Contoller_Core_Action{
 
 
     public function gridAction(){
-        $this->getMessage()->getSession()->start();
-        $query = "SELECT * FROM `category` WHERE 1";
-        $category = $this->getModel()->fetchAll($query);
         
-        $this->setCategory($category);
+        $this->getMessage()->getSession()->start();
+
+        try {
+            $query = "SELECT * FROM `category` WHERE 1";
+            $category = $this->getModel()->fetchAll($query);
+            if(!$category){
+                throw new Exception("data not found" ,1);
+            }
+            $this->setCategory($category);
+        } 
+        catch (Exception $e) {
+            $this->getMessage()->addMessages($e->getMessage(), Model_Core_Message::FAILURE);
+        }
 
         $this->getTemplate("category/grid.phtml");
-
     }
 
     public function addAction(){
@@ -57,52 +65,105 @@ class Controller_Category extends Contoller_Core_Action{
     }
 
     public function insertAction(){
+        $this->getMessage()->getSession()->start();
         $category = $this->getRequest()->getPost('category');
 
-        date_default_timezone_set('Asia/Kolkata');
-        $dateTime = date("Y-m-d h:i:sA");
-        $category['created_at'] = $dateTime;
-        
-        $this->getModel()->insert($category);
+        try {
+            if(!$category){
+                throw new Exception("data not inserted",1);
+            }
+            else{
+                $this->getMessage()->addMessages("category inserted successfully", Model_Core_Message::SUCCESS);
+            }
+            
+            date_default_timezone_set('Asia/Kolkata');
+            $dateTime = date("Y-m-d h:i:sA");
+            $category['created_at'] = $dateTime;
+            $this->getModel()->insert($category);
+        } 
+        catch (Exception $e) {
+            $this->getMessage()->addMessages($e->getMessage(), Model_Core_Message::FAILURE);
+        }
 
         $this->redirect('category', 'grid');
     }
 
     public function editAction(){
-
+        $this->getMessage()->getSession()->start();
         $query = "SELECT * FROM `category` WHERE `category_id` = '{$this->getRequest()->getParam('id')}'";
-
         $categoryRow = $this->getModel()->fetchRow($query);
 
-        $this->setCategory($categoryRow);
+        try {
+            if(!$categoryRow){
+                throw new Exception("row not found", 1);
+            }
+
+            $this->setCategory($categoryRow);
+        } 
+        catch (Exception $e) {
+            $this->getMessage()->addMessages($e->getMessage(), Model_Core_Message::FAILURE);
+        }
+
 
         $this->getTemplate("category/edit.phtml");
     }
 
     public function updateAction(){
+        $this->getMessage()->getSession()->start();
         $category = $this->getModel()->getPost('category');
 
-        date_default_timezone_set("Asia/Kolkata");
-        $dateTime = date("Y-m-d h:i:sA");
-        $category['updated_at'] = $dateTime;
+        try {
+            if(!$category){
+                throw new Exception("data not updated" , 1);
+            }
+            else{
+                $this->getMessage()->addMessages("data updated successfully" , Model_Core_Message::SUCCESS);
+            }
+
+            date_default_timezone_set("Asia/Kolkata");
+            $dateTime = date("Y-m-d h:i:sA");
+            $category['updated_at'] = $dateTime;
+        } 
+        catch (Exception $e) {
+            $this->getMessage()->addMessages($e->getMessage() , Model_Core_Message::FAILURE);
+        }
+
         
         $categoryId = $this->getRequest()->getParam('id');
 
-        $condition['category_id'] = $categoryId;
+        try {
+            if(!$categoryId){
+                throw new Exception("row id not found" , 1);
+            }
 
-        $this->getModel()->update($category,$condition);
+            $condition['category_id'] = $categoryId;
+            $this->getModel()->update($category,$condition);
+        } 
+        catch (Exception $e) {
+            $this->getMessage()->addMessages($e->getMessage(), Model_Core_Message::FAILURE);
+        }
 
-        // $this->redirect("index.php?c=category&a=grid");
         $this->redirect('category', 'grid' , [] , true);
-
     }
 
     public function deleteAction(){
+        $this->getMessage()->getSession()->start();
         $deleteId = $this->getRequest()->getParam('id');
+        
+        try {
+            if(!$deleteId){
+                throw new Exception("delete id not found", 1);
+            }
+            else{
+                $this->getMessage()->addMessages("data deleted successfully", Model_Core_Message::SUCCESS);
+            }
+            $this->getModel()->delete($deleteId);
 
-        $this->getModel()->delete($deleteId);
+        } 
+        catch (Exception $e) {
+            $this->getMessage()->addMessages($e->getMessage() , Model_Core_Message::FAILURE);
+        }
 
-        // $this->redirect("index.php?c=category&a=grid");
         $this->redirect('category', 'grid', [], true);
     }
 }
