+++ Controller/Customer.php
@@ -2,6 +2,7 @@
 require_once 'Model/Customer.php';
 require_once 'Controller/Core/Action.php';
 require_once 'Model/Core/Url.php';
+require_once 'Model/Core/Message.php';
 class Controller_Customer extends Contoller_Core_Action{
 
     protected $customer = [];
@@ -38,13 +39,22 @@ class Controller_Customer extends Contoller_Core_Action{
 
     
     public function gridAction(){
-         $query = "SELECT * FROM `customer` WHERE 1";
+        $this->getMessage()->getSession()->start();
+        $query = "SELECT * FROM `customer` WHERE 1";
         $customer = $this->getModel()->fetchAll($query);
-        
-        $this->setCustomer($customer);
+
+        try {
+            if(!$customer){
+                throw new Exception("data not found" , 1);
+            }
+
+            $this->setCustomer($customer);
+        } 
+        catch (Exception $e) {
+            $this->getMessage()->addMessages($e->getMessage(), Model_Core_Message::FAILURE);
+        }
 
         $this->getTemplate("customer/grid.phtml");
-        
     }
 
     public function addAction(){
@@ -52,52 +62,105 @@ class Controller_Customer extends Contoller_Core_Action{
     }
 
     public function insertAction(){
+        $this->getMessage()->getSession()->start();
         $customer = $this->getRequest()->getPost('customer');
-
-        date_default_timezone_set("Asia/Kolkata");
-        $dateTime = date("Y-m-d h:i:sA");
-        $customer['created_at'] = $dateTime;
         
-        $this->getModel()->insert($customer);
+        try {
+            if(!$customer){
+                throw new Exception("Data not inserted" , 1);
+            }
+            else{
+                $this->getMessage()->addMessages("customer added successfully" , Model_Core_Message::SUCCESS);
+            }
+
+            date_default_timezone_set("Asia/Kolkata");
+            $dateTime = date("Y-m-d h:i:sA");
+            $customer['created_at'] = $dateTime;
+            $this->getModel()->insert($customer);
+        } 
+        catch (Exception $e) {
+            $this->getMessage()->addMessages($e->getMessage() , Model_Core_Message::FAILURE);
+        }
 
         $this->redirect('customer', 'grid');
-
     }
 
     public function editAction(){
-
+        $this->getMessage()->getSession()->start();
         $query = "SELECT * FROM `customer` WHERE `customer_id` = '{$this->getRequest()->getParam('id')}'";
-
         $customerRow = $this->getModel()->fetchRow($query);
 
-        $this->setCustomer($customerRow);
+        try {
+            if(!$customerRow){
+                throw new Exception("row not found" , 1);
+            }
+
+            $this->setCustomer($customerRow);
+        } 
+        catch (Exception $e) {
+            $this->getMessage()->addMessages($e->getMessage() , Model_Core_Message::FAILURE);
+        }
 
         $this->getTemplate("customer/edit.phtml");
     }
 
     public function updateAction(){
+        $this->getMessage()->getSession()->start();
         $customer = $this->getRequest()->getPost('customer'); //$_POST > customer array
 
-        date_default_timezone_set("Asia/Kolkata");
-        $dateTime = date("Y-m-d h:i:sA");
-        $customer['updated_at'] = $dateTime;
+        try {
+            if(!$customer){
+                throw new Exception("customer row not found" , 1);
+            }
+            else{
+                $this->getMessage()->addMessages("data updated successfully" , Model_Core_Message::SUCCESS);
+            }
+
+            date_default_timezone_set("Asia/Kolkata");
+            $dateTime = date("Y-m-d h:i:sA");
+            $customer['updated_at'] = $dateTime;
+        } 
+        catch (Exception $e) {
+            $this->getMessage()->addMessages($e->getMessage() , Model_Core_Message::FAILURE);
+        }
+
 
         $customerId = $this->getRequest()->getParam('id');
 
-        $condition['customer_id'] = $customerId;
+        try {
+            if(!$customerId){
+                throw new Exception("custoemr id not found" , 1);
+            }
         
-        $this->getModel()->update($customer, $condition);
+            $condition['customer_id'] = $customerId;
+            $this->getModel()->update($customer, $condition);    
+        } 
+        catch (Exception $e) {
+            $this->getMessage()->addMessages($e->getMessage() , Model_Core_Message::FAILURE);
+        }
 
         $this->redirect('customer', 'grid' , [] , true);
     }
 
     public function deleteAction(){
+        $this->getMessage()->getSession()->start();
         $deleteId = $this->getRequest()->getParam('id');
 
-        $this->getModel()->delete($deleteId);
+        try {
+            if(!$deleteId){
+                throw new Exception("data id not found" , 1);
+            }
+            else{
+                $this->getMessage()->addMessages("data deleted successfully" , Model_Core_Message::SUCCESS);
+            }
+
+            $this->getModel()->delete($deleteId);
+        } 
+        catch (Exception $e) {
+            $this->getMessage()->addMessages($e->getMessage() , Model_Core_Message::FAILURE);
+        }
 
         $this->redirect('customer', 'grid' , [] , true);
-
     }
 }
 ?>
