+++ Controller/Payment.php
@@ -3,6 +3,7 @@
 require_once 'Controller/Core/Action.php';
 require_once 'Model/Payment.php';
 require_once 'Model/Core/Url.php';
+require_once 'Model/Core/Message.php';
 
 class Controller_Payment extends Contoller_Core_Action{
 
@@ -39,10 +40,20 @@ class Controller_Payment extends Contoller_Core_Action{
 
     //------------ all ACTIONS
     public function gridAction(){
+        $this->getMessage()->getSession()->start();
         $query = "SELECT * FROM `payment` WHERE 1";
         $payment = $this->getModel()->fetchAll($query);
-        
-        $this->setPayment($payment);
+
+        try {
+            if(!$payment){
+                throw new Exception("data not found",1);
+            }
+
+            $this->setPayment($payment);
+        } 
+        catch (Exception $e) {
+            $this->getMessage()->addMessages($e->getMessage() , Model_Core_Message::FAILURE)            ;
+        }
 
         $this->getTemplate("payment/grid.phtml");
     }
@@ -52,47 +63,104 @@ class Controller_Payment extends Contoller_Core_Action{
     }
 
     public function insertAction(){
+        $this->getMessage()->getSession()->start();
         $payment = $this->getRequest()->getPost('payment');
 
-        date_default_timezone_set("Asia/Kolkata");
-        $dateTime = date("Y-m-d h:i:sA");
-        $payment['created_at'] = $dateTime;
+        try {
+            if(!$payment){
+                throw new Exception("row not inserteD" , 1);
+            }
+            else{
+                $this->getMessage()->addMessages("payment row inserted successfully" , Model_Core_Message::SUCCESS);
+            }
 
-        $this->getModel()->insert($payment);
+            date_default_timezone_set("Asia/Kolkata");
+            $dateTime = date("Y-m-d h:i:sA");
+            $payment['created_at'] = $dateTime;
+            $this->getModel()->insert($payment);
+        } 
+        catch (Exception $e) {
+            $this->getMessage()->addMessages($e->getMessage() , Model_Core_Message::FAILURE);
+        }
 
         $this->redirect('payment' , 'grid');
     }
 
     public function editAction(){
+        $this->getMessage()->getSession()->start();
         $query = "SELECT * FROM `payment` WHERE `payment_method_id` = '{$this->getRequest()->getParam('id')}'";
-
         $paymentRow = $this->getModel()->fetchRow($query);
 
-        $this->setPayment($paymentRow);
+        try {
+            if(!$paymentRow){
+                throw new Exception("payment row not found" , 1);
+            }
+
+            $this->setPayment($paymentRow);
+        } 
+        catch (Exception $e) {
+            $this->getMessage()->addMessages($e->getMessage() , Model_Core_Message::FAILURE);
+        }
 
         $this->getTemplate("payment/edit.phtml");
     }
 
     public function updateAction(){
+        $this->getMessage()->getSession()->start();
         $payment = $this->getRequest()->getPost('payment');
 
-        date_default_timezone_set("Asia/Kolkata");
-        $dateTime =  date("Y-m-d h:i:sA");
-        $payment['updated_at'] = $dateTime;
+        try{
+            if(!$payment){
+                throw new Exception('data not found' , 1);
+            }
+            else{
+                $this->getMessage()->addMessages("data updated successfully" , Model_Core_Message::SUCCESS);
+            }
+
+            date_default_timezone_set("Asia/Kolkata");
+            $dateTime =  date("Y-m-d h:i:sA");
+            $payment['updated_at'] = $dateTime;
+        }
+        catch(Exception $e){
+            $this->getMessage()->addMessages($e->getMessage() , Model_Core_Message::FAILURE);
+        }
+
 
         $paymentId = $this->getRequest()->getParam('id');
 
-        $condition['payment_method_id'] = $paymentId;
+        try{
+            if(!$paymentId){
+                throw new Exception("payment id not found" ,1);
+            }
 
-        $this->getModel()->update($payment, $condition);
+            $condition['payment_method_id'] = $paymentId;
+            $this->getModel()->update($payment, $condition);
+        }
+        catch(Exception $e){
+            $this->getMessage()->addMessages($e->getMessage() , Model_Core_Message::FAILURE);
+        }
 
         $this->redirect('payment' , 'grid' ,[] , true);
     }
 
     public function deleteAction (){
+        $this->getMessage()->getSession()->start();
         $deleteId = $this->getRequest()->getParam('id');
 
-        $this->getModel()->delete($deleteId);
+        try{
+            if(!$deleteId){
+                throw new Exception("delte id not found" , 1);
+            }
+            else {
+                $this->getMessage()->addMessages("data deleted successfully" , Model_Core_Message::SUCCESS);
+            }
+            
+            $this->getModel()->delete($deleteId);
+        }
+
+        catch(Exception $e){
+            $this->getMessage()->addMessages($e->getMessage() , Model_Core_Message::FAILURE);
+        }
 
         $this->redirect('payment', 'grid', [], true );
     }
