+++ Controller/Shipping.php
@@ -1,10 +1,9 @@
 <?php
 
 require_once "Controller/Core/Action.php";
-
 require_once 'Model/Shipping.php';
 require_once 'Model/Core/Url.php';
-
+require_once 'Model/Core/Message.php';
 
 class Controller_Shipping extends Contoller_Core_Action{
 
@@ -43,13 +42,22 @@ class Controller_Shipping extends Contoller_Core_Action{
 
     // all actions
     public function gridAction(){
-         $query = "SELECT * FROM `shipping` WHERE 1";
+        $this->getMessage()->getSession()->start();
+        $query = "SELECT * FROM `shipping` WHERE 1";
         $shipping = $this->getModel()->fetchAll($query);
-        
-        $this->setShipping($shipping);
+       
+        try {
+            if(!$shipping){
+                throw new Exception("shipping data not found" , 1);
+            }
+
+            $this->setShipping($shipping);
+        } 
+        catch (Exception $e) {
+            $this->getMessage()->addMessages($e->getMessage() , Model_Core_Message::FAILURE);
+        }
 
         $this->getTemplate("shipping/grid.phtml");
-
     }
 
     public function addAction(){
@@ -57,47 +65,102 @@ class Controller_Shipping extends Contoller_Core_Action{
     }
 
     public function insertAction(){
+        $this->getMessage()->getSession()->start();
         $shipping = $this->getRequest()->getPost('shipping');
 
-        date_default_timezone_set("Asia/Kolkata");
-        $dateTime = date("Y-m-d h:i:sA");
-        $shipping['created_at'] = $dateTime;
+        try {
+            if(!$shipping){
+                throw new Exception("data not inserted" , 1);
+            }
+            else{
+                $this->getMessage()->addMessages("data inserted successfully" , Model_Core_Message::SUCCESS);
+            }
 
-        $this->getModel()->insert($shipping);
-        
+            date_default_timezone_set("Asia/Kolkata");
+            $dateTime = date("Y-m-d h:i:sA");
+            $shipping['created_at'] = $dateTime;
+            $this->getModel()->insert($shipping);
+        } 
+        catch (Exception $e) {
+            $this->getMessage()->addMessages($e->getMessage() , Model_Core_Message::FAILURE);
+        }
+
         $this->redirect('shipping', 'grid');
     }
 
     public function editAction(){
+        $this->getMessage()->getSession()->start();
         $query = "SELECT * FROM `shipping` WHERE `shipping_method_id` = '{$this->getRequest()->getParam('id')}'";
-
         $shippingRow = $this->getModel()->fetchRow($query);
 
-        $this->setShipping($shippingRow);
+        try{
+            if(!$shippingRow){
+                throw new Exception("shipping row not found" , 1);
+            }
+            
+            $this->setShipping($shippingRow);
+        }
+        catch(Exception $e){
+            $this->getMessage()->addMessages($e->getMessage() , Model_Core_Message::FAILURE);
+        }
 
         $this->getTemplate("shipping/edit.phtml");
     }
 
     public function updateAction(){
+        $this->getMessage()->getSession()->start();
         $shipping = $this->getRequest()->getPost('shipping');
 
-        date_default_timezone_set("Asia/Kolkata");
-        $dataTime = date("Y-m-d h:i:sA");
-        $shipping['updated_at'] = $dataTime;
+        try{
+            if(!$shipping){
+                throw new Exception("data not found" , 1);
+            }
+            else{
+                $this->getMessage()->addMessages("data updated success fully" , Model_Core_Message::SUCCESS);
+            }
+
+            date_default_timezone_set("Asia/Kolkata");
+            $dataTime = date("Y-m-d h:i:sA");
+            $shipping['updated_at'] = $dataTime;
+        }
+        catch(Exception $e){
+            $this->getMessage()->addMessages($e->getMessage() , Model_Core_Message::FAILURE);
+        }
+
 
         $shippingId = $this->getRequest()->getParam('id');
+        try{
+            if(!$shippingId){
+                throw new Exception("shipping id not found");
+            }
 
-        $condition['shipping_method_id'] = $shippingId;
-
-        $this->getModel()->update($shipping , $condition);
+            $condition['shipping_method_id'] = $shippingId;
+            $this->getModel()->update($shipping , $condition);
+        }
+        catch(Exception $e){
+            $this->getMessage()->addMessages($e->getMessage() , Model_Core_Message::FAILURE);
+        }
 
         $this->redirect('shipping', 'grid', [] , true);
     }
 
     public function deleteAction(){
+        $this->getMessage()->getSession()->start();
         $deleteId = $this->getRequest()->getParam('id');
-        
-        $this->getModel()->delete($deleteId);
+
+        try{
+            if(!$deleteId){
+                throw new Exception("Delete id not found" , 1);
+            }
+            else{
+                $this->getMessage()->addMessages("Row deleted successfully" , Model_Core_Message::SUCCESS);
+            }
+
+            $this->getModel()->delete($deleteId);
+        }
+        catch(Exception $e){
+            $this->getMessage()->addMessages($e->getMessage() , Model_Core_Message::FAILURE);
+        }
 
         $this->redirect('shipping', 'grid', [] , true);
     }
